chore: remove obsolete tests
