+++ tests/main_test.go
@@ -6,8 +6,6 @@ import (
 
 	"go-api-tests/internal/api"
 	"go-api-tests/internal/env"
-
-	"github.com/brianvoe/gofakeit/v6"
 )
 
 var (
@@ -17,12 +15,6 @@ var (
 	productionOrderEndpoint  string
 	productionOrdersEndpoint string
 )
-
-func init() { gofakeit.Seed(0) }
-
-func RandomNumber() int {
-	return gofakeit.Number(1, 10000)
-}
 
 func TestMain(m *testing.M) {
 	_ = env.LoadDotEnv(".env")
